Footprints in the Ash (#4484)
Ash will now track onto a user's boots, creating footprints in both the ash and any subsequent tiling the user steps on. Currently these footprints use the track template that blood uses, and indeed are a subtype of blood-tracks. The result is perhaps not the best sprite. Even so, I thought it at least a worthy experiment.