Shuriken Nerf (#4533)
Honestly I find that uranium ninja throwing stars are ridiculously overpowered and honestly just dumb in terms of mechanics. On top of the already ridiculous effects of getting hit by a throwing star, a uranium throwing star deals an additional random 25-50 toxin damage per hit regardless of armor, and regardless of whether or not it sticks in the wound. For reference, the xray rifle deals 25 damage per hit. I'm like 90% sure a uranium throwing star is more deadly than it's energy sword, which is just silly.

This update makes it so that the ninja throwing stars produced in the fabricator are of steel quality. To lessen the impact of this nerf, I also made some tweaks to uplink cost as well as the production cost of steel throwing star. Steel throwing stars also have less weight, and as an added bonus they can be purchased from the traitor uplink for 1 telecrystal each.