The Final Burrito Tweak to End All Burrito Tweaks (#4608)
After playing chef, I've discovered that there have been some new problems regarding burritos. I don't know why these are happening now as they weren't when I implemented it, so I'm changing recipes. For example:

Let's say that there are two recipies, one that calls for A, B, and C, and one that calls for A and B,
if I put in ingredients A, B, and C, it will make the recipe A and B.

I don't know why, but this PR makes it so that could never happen. Burritos are now made differently, and for the 100,000th time that nerd has to update the wiki again.