Adds the magic poppet (#4637)
This pr adds a magic poppet, a doll that can be infused with some spessmen's blood to bind them to it. From there, the person can mess with the dolls in some manners, and fun things will happen to the unfortunate victim of said sorcery. Added to some wizard spellbooks and a mining abandoned crate.