Reverts alcohol changes partially (#4722)
Restores old rates of gaining and losing intoxication.
The strength of alcohol was implemented with the original formula in mind, you cannot simply add another curve and expect it to fit. The rate of getting drunk was way too fast (increased 7 fold!). I also doubled the rate of losing alcohol, since gains and losses typically work in balance with one another.
Fixes a bug which spammed, "You're drunk"