Hostile AI fix (#4839)
When I made an update to hostile AI, I have not seen this stupid bug. This one variable makes them ignore people with full health. I urge to merge it ASAP as carps and spiders won't attack people with 100 health. fixes #4836

Also this makes hostile mobs to destroy surroundings if there is no-one to attack.