Air bubble (#4649)
This PR adds new item - airbubble

The idea is based of this thread.

Adds airbubble - item which when unfolded can provide user with environmental protection, the pressure inside won't be idea, but the oxygen will be enough to last for 20 minutes.
Airbubble properties:

When factory folded it is of normal size, so that it can fit in your internals box or backpack. But when folded manually after usage it is large size, and can only fit in storage like backpack minimal.

Has internal tank attached to it. By default each air bubble comes with double emergency oxygen tank(which can be found in engineering), it has enough oxygen for 20 minutes of usage. Tank can be detaches to be refilled and attach back. You can also attach different kind of tanks and they would work. Example is to attach phoron tank to kill someone, or sleep toxin to make them from breaking out.

Airbubble zipper can be forcefully blocked by wire handcuffs, to kidnap people. To take it out you need to use wirecutters.

Provides with breathable atmosphere, but pressure is just above minimal - meaning that you won't feel that comfortable inside of it

Breaking out when forcefully blocked is possible using resist, it would though make a hole in the bubble making it leak air of tank valve is on and not possible to use to protect yourself from environment.