Reduces the plasma cutter main rate (#4785)
What the title says, it was really high, having around 15% of instakilling someone on the first hit, if aiming at head, and 30% of chance of cutting a leg or arm at the first hit, this should reduce those chances. The main feature also takes in consideration the armor. So, plasma cutters are far more effective against less armored foes, limbs and xenos at all.