Mostly removes the stun mechanic from flashes, and other tweaks to flashes. (#4662)
The goal of this PR is to overall reduce the amount of oppressive non-violent stunlocking in the game without having to invest significant resources, force or effort into having to stun someone. Which starts with nerfing flashes a little.

Pepperspray isn't infinite and it's okay as it stands, especially seeing as how it can backfire. Flashes can still function as powerful utility. Flashbangs and bombs also benefit a bit from this buff against unprotected targets. Better wear flash protection!