Adds new disability, new chemical recipes. (#4727)
Adds a new special disability, love. Love works like monophobia, except that it specifically targets a single living mob and the owner must stay within the presence of that mob. The owner is also more susceptible to that mob's will.

Adds four new chemicals, with the potential for more:

1: Sovereign Glue
When applied to an object, the sovereign glue makes that object sticky. Anyone who picks it up cannot put it back down or have it taken from them unless the hand is severed or the glue removed. May consider other applications upon suggestion.

2: Universal solvent
Specifically counteracts sovereign glue. If more mundane applications can be suggested, this would probably not be a secret chemical. I can't think of any right now however.

3: Philter of Love
When ingested, the subject gains the love disability if they do not already have it.

4: Shapesand
Shapesand can be applied to any object. A counterfeit is created that looks and feels exactly like that item. If someone attempts to use the counterfeit, it collapses into sand (Specifically, in afterattack()).