Adds the gatling machine gun (#4867)
This pr adds a backpack mounted minigun. That requires the user to wear the power source on their back to deploy the weapon and fire it. It can supply nearly enough dakka, but it robs the user of wearing a backpack, a rig or carrying anything in their other hand while firing. It can be brought with around 50 telecrystals.

Sprites are from tg station.

/wetgnow