Attempts a fix for #5157.
For once, it wasn't actually Burger's fault, Burger did a completely reasonable thing and the ancient reagent code screwed it up. This should properly fix #5157, as well as a few possible future glitches in case we decide to add pouring out a specified number of units onto the floor, or mobs?