Surge prevention module balance (#5163)
So I was given input by the community that new IPC and borg surge modules are kinda useless. As they only provide 1-3 EMP protections given that ION rifles are 10 shots charge, and that IPC costs 14 TC which is more then half of traitor uplink. And borgs surge cost too much in resources to build. I expected that, so now I think I want to buff it.

IPC surge module now costs 12 TC(as opposed to 14).

comes in with 2-5 EMP protections which is RNG based.

Borg's surge module built cost in gold and silver was reduced by 50% from 10000 G and 15000 S to 5000 G and 7500 S.

Borg's surge module comes in with 2-5 EMP protections which is RNG based

Also I made surge's nanopaste to look fancier 💯