Adjusts Stun Baton Cranial Agony (#5239)
Adjusts the effect of a stun baton strike to the head. Instead of dealing 1.5x agony, it will now blur and disorientate the target.

Removes the arbitrary limitation on most hats that prevented them from actually covering the head. Their status as non-combat items should be represented by other values, as is standard for the rest of the clothing items.

Tweaks the head organ to also include the FACE as a body part applicable to def_zones. This may have some other trickle down effects, but should work normally.