Adds ouija boards and magic eightballs (#5238)
Both items taken from the tgstation.

Magic eightballs allow you to ask a question, picking up a random answer. Or they might ask the dead on the deadchat, if it is a haunted one.

The ouija board allows some possible limited communication with the ghosts, however anyone else can move the planchette, allowing people to mess with other person as well.