Adds proper power and atmospherics pipes to the construction level.
Ran pipes and power through maintenance by the service ladders through the cargo interstitial level, out the airlock into space, and back up into the airlock of the construction level. Added a new area for the construction level specifically so it's not tied to surface maintenance. Made the airlocks larger for miners that go out the construction level, and for engineers that need to carry things out.