Lighter tweaks (#5479)
Zippo messages are no longer red. For those who want to opt out of the zippo flavor text spam, but still have a lighter at loadout, the cheap lighter has been added in there. The zippo's description has been tweaked so that it's slightly more obvious that it's going to spam flavor messages.