Gamemode Requirement Tweaks (#5247)
A lot of the gamemodes had bad requirements for players and antags, and as a result, it would result in bad rounds that were either incredibly boring or one sided. Gamemodes would appear to be far too common, such as changeling and vampire, as the requirements for players and antags are VERY little.