Tesla Movement Reworks (#5693)
The Tesla has not been nerfed or buffed, but more appropriately reworked.

The tesla can move across z-levels when it is not contained, however it cannot phase through reinforced floors or walls. These new features mostly contain the tesla to the engineering sublevel for about 10 minutes, until RNG dictates that it reaches other departments on all levels, or if very lucky, escapes to the derelict and doesn't bother the station for another 30 minutes.