Adds Steamed Crab Legs (#5869)
Allows chefs to make "steamed crab legs" which is just plain tasty.

Crabs drop crab legs when butchered (testing shows they drop 3 from the gibber).
10u water, 1 spread, 1u salt, 1 crab legs makes steamed crab legs

I made a box of 5 crab legs (in a similar fashion to the clams) that are meant to be orderable from cargo. Crabs aren't exactly common on Aurora, after all.

Sprites for the legs and the dish are thanks to Meowy with minor photoshop editing from myself.