Small Bugfixes (#5915)
When you click the security records console with an ID, it no longer gives you the violent message of you hitting it. You just put the ID in, like a normal person.

You can put your ID directly into the ore processing console (like you can with every other ID console) instead of hitting it with your ID like a spaz and then having to go into the interface and click the "Insert ID" line to put your ID in. (The line is still an option, like with other consoles.)

The RIG module press lets you put your to-be-pressed item inside without having to hit it with the item along the way.