Knives now eyestab (#5922)
All knives (including the trench knife, which was previously a sword, and now has lost its ridiculously low chance to parry) have been moved to the knife datum, which are defined by their ability to eyestab, and be concealed in boots. The butcher's cleaver and the meat hook are no longer considered knives - the cleaver is now a hatchet, and the hook is a regular weapon.
The tactical knife has been buffed, and deals 18 damage, to be superior/equivalent in power to the steel bat, considering that it's an illegal/hidden item.