Fixing Z level cable update and substations (#5937)
Oh boi, this one is a tiny in terms of lines, yet a huge impact. It absolutely solves problem with powernet bug of cables not updating through Z levels properly. Spend lots of time and especially intellectual effort to debug this thing and finally comparing Bay's code and debugging ours showed that lines 359 and 364 should be power_list(T, src, 12) because that stands for marked and we have unmarked cables.

Fixes Z level update of cables, now it updates them properly. So the substations work. Fixes #5165 , Fixes #5883 , Fixes #5655

Fixes mapping of cables

Increases charge of runtime SMES

Adds Z powernet to runtime levels for debug and tests

Adds Z level supply pipes and vents for debugging and tests