Buffs ore summoners (#5849)
Allow me to explain in the event someone potentially complains about this.

Each vein of ore in a single tile drops roughly 5 bits of that ore type. It doesn't matter if it's diamonds, coal or whatever else.

The problem with this, however, is that the ore summoner is only adjusted to grab 10 bits of ore and nothing else, making it inefficient and useless to buy in the mining vendor.

This was changed because of 'performance problems' (whether this is true I have no idea) but only 10 bits of ore is just ridiculous in my opinion, because it means it'll only draw from 2 veins mined out, which is not effective in the slightest. You'd be better off using the ore magnet right now.

Of course, this change hopes to change that so that this isn't a useless item to buy from the ore vendor and to (somewhat) restore it to its former glory. Albeit with a sanity check.