Replaces the tranquilizer rifle magazine with a box (#5952)
The head of security officer came with a magazine for the tranquilizer rifle, however the rifle did not accept magazines, and worked like the regular am rifle, where you have to load one shell at the time. This pr replaces the magazine, that forced you to empty it on the ground to load the rifle, with a regular box full of darts.