Hoodie rework & resprite (#5998)
Hoodies have been resprited to be less ugly, and now also function in a similar way to winter coats, i.e. actually having a hood. In addition to that, they can still be opened and closed. They're also recolorable in the loadout.
The hood icon not being colored is a known issue, but I'm not sure how to fix it. This affects roughly nothing except for the UI.
The PR also makes a slight tweak to my cardigan sprite for it to be less ugly.