Power Insulated (#6055)
It is now no longer to electrocute yourself by merely clicking on an exposed power cable with an object that happens to be conductive. Power cables are now insulated.

It is now possible to sever cables with an object that is sharp or edgy by clicking on an exposed power cable with said object while on help intent (to alleviate potential misclicks while fighting in maint, which one assumes is normally done in harm intent.) If the sharp object is conductive you will electrocute yourself.