Hostile mobs line of fire check (#6075)
This PR make few things:

Hostile mobs now have a var smart which designates if they are smart enough to check their line of fire. In future I will also use this var to try and find path to their target around obstacles.

If Hostile mob is smart and is a ranged mob, it will check it's line of fire before it shoots. If there is a friendly mob in line of fire, we will not shoot. If we hit some other object, but we still can hit our target penetrating that object, we will fire. If we can't hit our target, we just don't fire.

Currently makes all hostile, ranged mobs smart.

Cleans up hivebot code and improves it. Hivebot spawn beacon has now chance of spawning different kinds of bots each cycle, and has a spawn delay between each.