Fixing Jump To Mob(And admin proc too) (#6140)
So I noticed that Jump To Mob proc of both observer and admin version is using wrong list, and it does not actually work on any mob because the mob it is looking for is of different format in the list.

Fixes Jump To Mob for Ghost tab and Admin tab.

Makes Jump To Mob use same list as Follow. Which makes it list things like carp(2) or if mob is dead or a ghost.

Renames admin Jump To mob. Making admin panel use admin proc, instead of public.