Timers sanity checks Port from TG. (#6125)
This adds timers sanity checks with stack tracing of where it was called. Check are:

If timer had no callback(crashes timer)

If timer was called with negative wait time. Converts time to 0 and gives call stack.

If timer was called on deleted or about to be deleted object. Still runs, but gives call stack

If timer had infinite or more wait time. Crashes timer

Adds sanity check to beams for timer wait time.