Lazy (possible) fix for anesthesia (#6199)
So while observing a bit, I noticed something. The "sleeping" variable of the mob, which controls whether or not the mob is awake (anything > 0 means that the mob is asleep) gets reduced at quite a rapid pace. And would often reach a value of 1 during regular gameplay. This is kinda flying close to the sun.