A new skrellian plant, and associated stuff (#6340)
Adds dyn, a Skrellian herb. Renowned as a medicine, its juice has mild anti-toxin properties identical to how powerful tea was before Burger nerfed it; lower than tea because Arrow considered old tea to be too OP, and any beverages made from it do too. Seeds are available both in the garden and hydroponics, and the plant may occasionally contain actual dylovene in it.
The new drinks are the raw juice of the herb, a tea (1 juice to 4 water), and an icy/fizzy iced tea thing (1 juice to 2 ice to 2 soda water). The last comes as a soda can in the vendor, and a carton of juice is available at the bartender's boozemat.
This is the first in a series of PRs expanding upon skrellian cuisine/drinks.