The great hydroponics belt (#6347)
REASON

I bloody hate how I have to throw all my crap into my backpack and I don't have a belt that can store my crap onto my belt slot. This aims to have a hydroponics belt be added via the lockers and bio-thingy.

Also, it'll add a belt that defines hydroponics, and has a few things added to it so it gives you a reason to use it instead of the regular utility belt.

Also, I'm removing the chemical-grenade restriction to cleaning from the janitor belt. Now it can hold all grenades, same with this new belt.