Fixes Raiders Hazard Rig (#6441)
Per #6429, someone mapped in a hazard suit into the pirate ship with a modified (and invalid) newlist access value that ATLAS could not understand.

To fix this, I have created a new pirate-only subtype of the hazard rig that can be accessed with the pirate's syndicate IDs. I also removed its ability to be viewed through station cameras at the same time, as an added bonus.