Changes fireaxe to only cleave when wielded (#6548)
As an engineering main that had to make use of the fireaxe to deal with biohazards, it got old fast when someone entered your range while you were swinging and you decapitate them with the might of Zeus. That can still happen, but, if you know you're gonna be swinging when people are around, you can do it with only one hand to prevent multi-decapitation.

Besides, how well can you drive a fireaxe through multiple bodies if you're only using one arm?

Sidenote: I wanted to do this with other weapons as well, but my brain is simply too small to comprehend what the hell's going on with half of them, so, I'm just sticking with the fireaxe for now.