Fixes autohiss breaking for humans (#6634)
Humans could enable autohiss. This resulted in the handle_autohiss proc runtiming, due to missing expected maps. The runtiming rendered the person unable to speak until they turned autohiss off again.

Fixes this by checking for the presence and length of autohiss_basic_map. The map is expected as a predicate towards autohiss working in the first place.