Adds the ability to lay out towels flat on the ground (#6631)
After seeing in the recent beach episode that players had to deal with only being able to relax on folded up towels, I've added a verb to them that lays it out flat on the ground instead. When you pick up the towel it turns back into its regular self again.