Prevents ghosts from freely observing on the derelict.
Ghosts can now only observe on the derelict by following a member of a away mission.

If they stop following a member of they away mission, they are teleported back to the station.

They are also forced into the vision mode of a normal human mob as long as they are on the derelict.