Arrow768 2 hours ago  Member Now that I've actually spawned it we might want to ban both from traders because under the id the game expects the scanner, instruction booklet AND a wrapped labeled package which spawns code to the machine upon unwrapping.
I dont think that is correct, as the actual items exist under /obj/item/device/eftpos and /obj/item/device/nanoquikpay and only during spawn the paper and the box is created, but that doesnt change the path of the item itself and it should still be sellable.

That said, I dont have an issue with blacklisting both because a merchant would have a hard time acquiring either one of those.