Liver Tweaks (#6874)
The liver's ability to deal with toxins was super ultra giga-bad. This lead to damage that was pretty annoying to deal with, and definite round-enders for Vaurca.
This PR adds a tolerance level to livers, and gives Vaurca and Vox livers a higher tolerance to toxins, that being up to 20 toxins. It also minorly boosts all the other species' tolerance to 5 toxins, rather than 3.
Vaurca get the toxin chemical added to their bloodstream when they breathe in oxygen, so damage will persist if you start breathing phoron after having no oxygen for a bit. Get to medical ASAP if this happens, anti-toxin will nullify the liver damaging process completely.
With this update, you have slightly more leeway with lag and mask switches.