Machine Overload Nerf (#7025)
Machine overloads are now no longer a joke. The strength of APC machine overloads is now reduced by roughly 2/3rds. As you can see by the below graph, this change is very much needed.

image

X axis is the APC's power cell charge.
Y axis is the strength of the explosion.

Purple line is the strength of the explosion in its current state. As you can see above, it caps out very easily and super fast: Basically 4 AA batteries can destroy nearly everything.

Red line is the new strength of the explosion in this PR. It's much more reasonable and caps out at APCs with upgrades.

That purple dotted line down the middle is the cell capacity of a normal APC.