Changeling Arm only functions as crowbar with help intent active (#7481)
Fixes #7477

I tested this with attacking a robot and a regular wall with help intent on/off. It seems there's something custom with walls for changeling so please consider any side effects of changing the iscrowbar logic.

One thing is, even with help intent you will attack humans when clicking them. While this fix allows robots to be attacked by the changeling arm, it can not do so on help, which may seem inconsistent.