Station Safe bugfix (#7602)
Fixes a bug where station safe could be impossible to open if certain tumbler_open values were rolled.

The bug in question is caused by tumbler rolling which is a core process of the safe-cracking minigame.

After testing it on private server, I have determined that changing tumbler rolling values creates way more newer issues, and as such solved the rare glitch by making sure that any and all values rolled by a safe on initiation are possible to be solved within confines of tumbler rolling

[35 differential value between tumbler_open 1 and 2] = Max of tumbler rolling.

Station Safes now check on initiation if the differential is bigger than 34. If yes, they reduce the differential to 34, making the otherwise impossible station safe spawn possible to crack.